Refactor InlinedsPositioner into InlinedSourceMap
